nueva arquitectura
base de la nueva arquitectura
se mantienen los archivos py originales como referencia
