+++ frontend/static/css/style_create-character.css
