+++ frontend/static/scripts/attributeLogic.js
