+++ frontend/static/scripts/characterDataProcess.js
