+++ frontend/static/scripts/hpLogic.js
