+++ frontend/static/scripts/loadItemLanguage.js
