+++ frontend/static/scripts/loadSkills.js
