+++ frontend/static/scripts/sendNewCharacterData.js
