+++ frontend/templates/auth/register.html
