+++ frontend/templates/home.html
@@ -2,6 +2,9 @@
 {% block title %} DnD-WBM {% endblock %}
 {% block body %}
 <main>
+    {% if data %}
+        <p>{{data}}</p>
+    {% endif %}
     <div class="learnGame">
         <section>
             <h2 class="textShadow">New to D&D?</h2>
